admin_api/requirements.txt
Django
djangorestframework 
python-decouple
django-cors-headers
pika
gunicorn
psycopg2-binary
celery